[Refactor] Picker관련 코드 개선

+++ MateRunner/MateRunner/Data/Repository/Common/DefaultUserRepository.swift
@@ -53,11 +53,11 @@ final class DefaultUserRepository: UserRepository {
         return self.networkService.documentDoesExist(collection: FirebaseCollection.user, document: nickname)
     }
     
-    func saveUserInfo(uid: String, nickname: String, height: Int, weight: Int) -> Observable<Bool> {
+    func saveUserInfo(uid: String, nickname: String, height: Double, weight: Double) -> Observable<Bool> {
         let data: [String: Any] = [
             "nickname": nickname,
-            "height": Double(height),
-            "weight": Double(weight),
+            "height": height,
+            "weight": weight,
             "time": 0,
             "distance": 0.0,
             "calorie": 0.0,
